API Move serpentTools/objects/readers to s/o/base.py
