+++ serpentTools/objects/base.py
