+++ serpentTools/parsers/bumat.py
@@ -1,6 +1,6 @@
 """Parser responsible for reading the ``*bumat<n>.m`` files"""
 
-from serpentTools.objects.readers import MaterialReader
+from serpentTools.objects.base import MaterialReader
 
 
 class BumatReader(MaterialReader):
